refactor: move all objects into separate packages according to the role they play within the program
