+++ src/main/java/sos/ryanbyers/gameLogic/SOSGame.java
@@ -1,0 +1,4 @@
+package sos.ryanbyers.gameLogic;
+
+public interface SOSGame {
+}
