+++ src/main/java/sos/ryanbyers/gameLogic/SOSGeneralGame.java
@@ -1,0 +1,4 @@
+package sos.ryanbyers.gameLogic;
+
+public class SOSGeneralGame {
+}
