+++ src/main/java/sos/ryanbyers/gameLogic/TurnManager.java
@@ -1,4 +1,4 @@
-package sos.ryanbyers.sosjavafx;
+package sos.ryanbyers.gameLogic;
 
 import javafx.scene.control.Alert;
