+++ src/main/java/sos/ryanbyers/gui/ButtonHolder.java
@@ -1,4 +1,4 @@
-package sos.ryanbyers.sosjavafx;
+package sos.ryanbyers.gui;
 
 import javafx.scene.control.Button;
 import javafx.scene.control.RadioButton;
