+++ src/main/java/sos/ryanbyers/input/BoardButtonListener.java
@@ -1,4 +1,4 @@
-package sos.ryanbyers.sosjavafx;
+package sos.ryanbyers.input;
 
 public class BoardButtonListener {
 }
